delete source format endpoint implementation

+++ src/main/java/com/encyclopediagalactica/sourceformats/services/implementations/DeleteByIdService.java
@@ -3,20 +3,24 @@ package com.encyclopediagalactica.sourceformats.services.implementations;
 import com.encyclopediagalactica.sourceformats.repositories.SourceFormatRepository;
 import com.encyclopediagalactica.sourceformats.services.interfaces.DeleteByIdServiceInterface;
 import lombok.NonNull;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Service;
 
 @Service
 public class DeleteByIdService implements DeleteByIdServiceInterface {
-  
+
   private final SourceFormatRepository sourceFormatRepository;
 
-  public DeleteByIdService(
-      @NonNull SourceFormatRepository sourceFormatRepository) {
+  public DeleteByIdService(@NonNull SourceFormatRepository sourceFormatRepository) {
     this.sourceFormatRepository = sourceFormatRepository;
   }
 
   @Override
   public void deleteById(Long id) {
+
+    if (id == 0) {
+      throw new InvalidDataAccessApiUsageException("The provided entity Id cannot be 0");
+    }
     this.sourceFormatRepository.deleteById(id);
   }
 }
